requirements.txt
Flask==3.1.2
lxml==6.0.2
requests==2.32.5
python-dotenv==1.1.1